move everything to its folder
