+++ src/components/tableV2/components/comparisonFilter.js
